moved remind functionality to new class so that there can be multiple reminders at once
